feat: moves getAddressesFromGenesis into accounts.go from test
