+++ internal/accounts.go
@@ -2,8 +2,11 @@ package internal
 
 import (
 	"context"
+	"encoding/json"
 	"errors"
 	"fmt"
+	"io"
+	"net/http"
 	"time"
 
 	"github.com/algorandfoundation/hack-tui/api"
@@ -25,6 +28,80 @@ type Account struct {
 	Expires time.Time
 	// The LastModified round, this only pertains to keys that can be updated
 	LastModified int
+}
+
+func getAddressesFromGenesis() ([]string, string, string, error) {
+
+	// TODO: replace with calls to GetGenesis
+
+	resp, err := http.Get("http://localhost:8080/genesis")
+	if err != nil {
+		return []string{}, "", "", err
+	}
+
+	if resp.StatusCode != 200 {
+		return []string{}, "", "", errors.New(fmt.Sprintf("Failed to get genesis file. Received error code: %d", resp.StatusCode))
+	}
+
+	defer resp.Body.Close()
+
+	// Read the response body
+	body, err := io.ReadAll(resp.Body)
+	if err != nil {
+		return []string{}, "", "", err
+	}
+
+	// Unmarshal the JSON response into a map
+	var jsonResponse map[string]interface{}
+	err = json.Unmarshal(body, &jsonResponse)
+	if err != nil {
+		return []string{}, "", "", err
+	}
+
+	// Two special addresses
+	rewardsPool := "7777777777777777777777777777777777777777777777777774MSJUVU"
+	feeSink := "A7NMWS3NT3IUDMLVO26ULGXGIIOUQ3ND2TXSER6EBGRZNOBOUIQXHIBGDE"
+	rewardsPoolIncluded := false
+	feeSinkIncluded := false
+
+	// Loop over each entry in the "alloc" list and collect the "addr" values
+	var addresses []string
+	if allocList, ok := jsonResponse["alloc"].([]interface{}); ok {
+		for _, entry := range allocList {
+			if entryMap, ok := entry.(map[string]interface{}); ok {
+				if addr, ok := entryMap["addr"].(string); ok {
+					if addr == rewardsPool {
+						rewardsPoolIncluded = true
+					} else if addr == feeSink {
+						feeSinkIncluded = true
+					} else {
+						addresses = append(addresses, addr)
+					}
+				} else {
+					return []string{}, "", "", fmt.Errorf("In genesis.json no addr string found in list element entry:  %+v", entry)
+				}
+			} else {
+				return []string{}, "", "", fmt.Errorf("In genesis.json list element of alloc-field is not a map:  %+v", entry)
+			}
+		}
+	} else {
+		return []string{}, "", "", errors.New("alloc is not a list")
+	}
+
+	if !rewardsPoolIncluded || !feeSinkIncluded {
+		return []string{}, "", "", errors.New("Expected RewardsPool and/or FeeSink addresses NOT found in genesis file")
+	}
+
+	return addresses, rewardsPool, feeSink, nil
+}
+
+func isValidStatus(status string) bool {
+	validStatuses := map[string]bool{
+		"Online":            true,
+		"Offline":           true,
+		"Not Participating": true,
+	}
+	return validStatuses[status]
 }
 
 // Get Online Status of Account
